Crud for Owner

+++ app/Listeners/SendOwnerCreatedNotification.php
@@ -5,6 +5,7 @@ namespace App\Listeners;
 use App\Events\OwnerCreated;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Support\Facades\App;
 
 class SendOwnerCreatedNotification
 {
@@ -26,9 +27,11 @@ class SendOwnerCreatedNotification
      */
     public function handle(OwnerCreated $event)
     {
-        //
-		//dd($event->owner->id); //created owner
-		$event->owner->forceDelete();
-		dd("Event Listener works. Created Owner {$event->owner->id} was deleted at once (App\Events\OwnerCreated <==> App\Listeners\SendOwnerCreatedNotification)");
-    }
+		
+        if(App::runningInConsole()){ //run only if triggered in console (routes/console) (php artisan event-listener:start), so not to delete real created owners
+		    //dd($event->owner->id); //created owner
+		    $event->owner->forceDelete();
+		    dd("Event Listener works. Created Owner {$event->owner->id} was deleted at once (App\Events\OwnerCreated <==> App\Listeners\SendOwnerCreatedNotification)");
+        } 
+	}
 }
